refactor(crewai): group tests by feature, share build exclude, trim pyproject comments (CPK-7716 review)
