+++ integrations/crew-ai/python/pyproject.toml
@@ -9,11 +9,7 @@ authors = [
 ]
 readme = "README.md"
 requires-python = ">=3.10,<3.14"
-# Runtime deps are a faithful, semantics-preserving translation of the
-# previous Poetry caret ranges (CPK-7716 Poetry -> uv migration). Poetry's
-# ``^X.Y.Z`` on a 0.x version caps at the next minor, so ``^0.130.0`` is
-# ``>=0.130.0,<0.131.0`` etc. The crewai pin is preserved EXACTLY: the
-# crewai 1.x bump is a separate ticket (CPK-7718) — do not widen it here.
+# Poetry caret ranges translated to explicit bounds; crewai pin preserved for CPK-7718.
 dependencies = [
     "ag-ui-protocol>=0.1.10",
     "fastapi>=0.115.12,<0.116.0",
@@ -28,13 +24,7 @@ dev = "ag_ui_crewai.dojo:main"
 dev = [
     "pytest>=8.0,<9.0",
     "pytest-asyncio>=0.24,<0.25",
-    # Dev-only dep. Lower bound at 0.27 so transitive deps (fastapi test
-    # client) continue to resolve to the current 0.28.x branch rather than
-    # being accidentally downgraded. Upper bound <0.29 (finding #21) so a
-    # breaking 0.x minor release (httpx is pre-1.0 and has historically
-    # shipped transport-API breaks at minor-version boundaries) cannot
-    # silently land in CI. Bump this ceiling deliberately when verified
-    # against the new minor.
+    # Dev-only; <0.29 guards against a breaking httpx 0.x minor.
     "httpx>=0.27,<0.29",
 ]
 
@@ -42,28 +32,23 @@ dev = [
 requires = ["hatchling"]
 build-backend = "hatchling.build"
 
-[tool.hatch.build.targets.wheel]
-packages = ["ag_ui_crewai"]
+# Shared across wheel and sdist so the exclude list is not repeated per target.
+[tool.hatch.build]
 exclude = [
     "ag_ui_crewai/dojo.py",
     "ag_ui_crewai/examples/**",
 ]
 
-# Keep the source archive lean, matching the pre-migration Poetry sdist
-# (~31 KB). Hatch's default sdist would otherwise sweep in tests, uv.lock,
-# .gitignore, dojo.py and examples (~283 KB). Ship only the importable
-# package plus packaging metadata, and drop the same non-shipping modules
-# the wheel excludes.
+[tool.hatch.build.targets.wheel]
+packages = ["ag_ui_crewai"]
+
+# Lean sdist: package + metadata only (~31 KB), matching the pre-migration Poetry sdist.
 [tool.hatch.build.targets.sdist]
 include = [
     "ag_ui_crewai",
     "README.md",
     "LICENSE",
     "pyproject.toml",
-]
-exclude = [
-    "ag_ui_crewai/dojo.py",
-    "ag_ui_crewai/examples/**",
 ]
 
 [tool.pytest.ini_options]
